Trait kuehree::sum_query::IndexableSumQuery

pub trait IndexableSumQuery<T> where
    T: Copy + Sub<Output = T> + Add<Output = T>, {
    type PrefixSumContainer: IntoIterator<Item = T> + Index<usize, Output = T>;
    fn new(data: impl IntoIterator<Item = T>) -> Self;

    fn prefix_sum_array(&self) -> &Self::PrefixSumContainer;

    fn query(&self, start: usize, end: usize) -> T { ... }

    fn non_zero_query(&self, start: NonZeroUsize, end: NonZeroUsize) -> T { ... }
}

This trait provides methods required for SumQuery types

Associated Types

type PrefixSumContainer: IntoIterator<Item = T> + Index<usize, Output = T>

Required methods

fn new(data: impl IntoIterator<Item = T>) -> Self

Construct Self

Algorithmic complexity: O(n)

fn prefix_sum_array(&self) -> &Self::PrefixSumContainer

Provided methods

fn query(&self, start: usize, end: usize) -> T

Query between start range and end range

Negative querying is not implemented, hence end has to be greater or equal to start

fn non_zero_query(&self, start: NonZeroUsize, end: NonZeroUsize) -> T

Query between start range and end rage

This function elides the branch for when start == 0